test: share WI15 test fakes and evidence builders
Move the copy-pasted scaffolding (script-test harness, Motion device
and clock fakes, list-recording recorder, Ophir setting-evidence tuple,
safety config / power-cycle / measurement builders, FPGA revision
constant) into wi15_fakes, wi15_builders, and wi15_script_harness.
Fakes that genuinely differ between files stay where they are.
