+++ tests/test_wi15_safety_calibration_report.py
@@ -4,7 +4,6 @@ from omotion.calibration.laser import (
     CriterionResult,
     DeviceIdentity,
     FailureKind,
-    FpgaFirmwareRevision,
     FinalSettingCheck,
     ProcedureStatus,
     SettingReadback,
@@ -14,7 +13,6 @@ from omotion.calibration.safety import (
     ADC_ROUNDING_RULE,
     AdcReadEvidence,
     NormalScanEvidence,
-    PowerCycleEvidence,
     PulseLimitCalculation,
     SafetyLimitCalculation,
     SafetyWarningEvidence,
@@ -26,6 +24,7 @@ from omotion.calibration.safety_workflow import (
     SafetyCalibrationResult,
 )
 from omotion.calibration.single_sensor_laser import ProcedureEvent
+from wi15_builders import FPGA_REVISIONS, valid_power_cycle, valid_safety_config
 
 
 NOW = datetime(2026, 8, 13, 16, 0, tzinfo=timezone.utc)
@@ -58,18 +57,7 @@ def _warning(*, ok=True, faults=()):
 
 
 def _current_config():
-    return {
-        "TA_PULSE_WIDTH": 500,
-        "TA_CURRENT_DRV": 5000,
-        "SEED_CW_GAIN": 140,
-        "EE_PULSE_WIDTH_UL": 550,
-        "EE_RATE_LL": 23125,
-        "EE_DRIVE_CL": 9999,
-        "OPT_PULSE_WIDTH_UL": 550,
-        "OPT_RATE_LL": 23125,
-        "OPT_DRIVE_CL": 9999,
-        "TEC_TRIP": 40,
-    }
+    return valid_safety_config()
 
 
 def _result(**changes):
@@ -105,20 +93,7 @@ def _result(**changes):
         550,
         "derived",
     )
-    power_cycle = PowerCycleEvidence(
-        NOW,
-        NOW + timedelta(seconds=1),
-        NOW + timedelta(seconds=16),
-        NOW + timedelta(seconds=16),
-        NOW + timedelta(seconds=20),
-        15.0,
-        True,
-        True,
-        True,
-        "same Motion handle disconnected and reconnected",
-        "C-1",
-        "C-1",
-    )
+    power_cycle = valid_power_cycle(NOW)
     scan = NormalScanEvidence(
         ShippingTopology.SINGLE_LEFT,
         30.0,
@@ -149,10 +124,7 @@ def _result(**changes):
             "console-fw",
             "console-hw",
             "camera-fpga-omitted",
-            tuple(
-                FpgaFirmwareRevision(controller, "1.2.3")
-                for controller in ("TA", "SEED", "SAFETY_EE", "SAFETY_OPT")
-            ),
+            FPGA_REVISIONS,
         ),
         "current_configuration": current,
         "configuration_criteria": (
