+++ tests/test_wi15_single_sensor_laser_script.py
@@ -1,6 +1,4 @@
-import importlib.util
 import json
-import sys
 from dataclasses import dataclass
 from datetime import datetime, timezone
 from pathlib import Path
@@ -13,25 +11,22 @@ from omotion.calibration.single_sensor_laser import (
     ReportArtifactStatus,
     SingleSensorLaserCalibrationResult,
 )
-
-
-SCRIPT_PATH = (
-    Path(__file__).resolve().parents[1]
-    / "scripts"
-    / "wi15_single_sensor_laser_calibration.py"
+from wi15_script_harness import (
+    FakeBench,
+    FakeMeter,
+    FakeRecorder,
+    FakeReport,
+    complete_args,
+    load_wi15_script,
+    wi15_script_path,
 )
 
 
+SCRIPT_PATH = wi15_script_path("wi15_single_sensor_laser_calibration.py")
+
+
 def load_script():
-    spec = importlib.util.spec_from_file_location("wi15_script_under_test", SCRIPT_PATH)
-    module = importlib.util.module_from_spec(spec)
-    assert spec.loader is not None
-    sys.modules[spec.name] = module
-    try:
-        spec.loader.exec_module(module)
-    finally:
-        sys.modules.pop(spec.name, None)
-    return module
+    return load_wi15_script(SCRIPT_PATH, "wi15_script_under_test")
 
 
 @dataclass(frozen=True)
@@ -43,34 +38,6 @@ class FakeResult:
     report_artifact: object | None = None
 
 
-class FakeRecorder:
-    def __init__(self, root):
-        self.run_directory = Path(root) / "run"
-        self.run_directory.mkdir(parents=True)
-        self.json_path = self.run_directory / "run.json"
-        self.checkpoints = []
-
-    def checkpoint(self, result):
-        self.checkpoints.append(result)
-
-
-class FakeMeter:
-    def __init__(self):
-        self.closed = 0
-
-    def close(self):
-        self.closed += 1
-
-
-class FakeBench:
-    def __init__(self, meter):
-        self.meter = meter
-        self.closed = 0
-
-    def close(self):
-        self.closed += 1
-
-
 class FakeWorkflow:
     def __init__(self, result):
         self.result = result
@@ -79,17 +46,6 @@ class FakeWorkflow:
     def run(self, request):
         self.requests.append(request)
         return self.result
-
-
-class FakeReport:
-    def __init__(self, directory):
-        self.report_path = Path(directory) / "report.html"
-        self.writes = []
-
-    def write(self, request, result, json_path):
-        self.writes.append((request, result, Path(json_path)))
-        self.report_path.write_text("report", encoding="ascii")
-        return self.report_path
 
 
 def configured_script(monkeypatch, tmp_path, result=FakeResult(ProcedureStatus.PASSED)):
@@ -105,21 +61,6 @@ def configured_script(monkeypatch, tmp_path, result=FakeResult(ProcedureStatus.P
     monkeypatch.setattr(script, "workflow_factory", lambda *args: workflow)
     monkeypatch.setattr(script, "report_factory", lambda directory: report)
     return script, recorder, meter, bench, workflow, report
-
-
-def complete_args(tmp_path):
-    return [
-        "--output-dir",
-        str(tmp_path),
-        "--operator",
-        "operator",
-        "--build-revision",
-        "build-7",
-        "--fixture-id",
-        "fixture-2",
-        "--fixture-calibration-status",
-        "current",
-    ]
 
 
 def answers(*values):
